rendering: Rect's x and y have to be signed

+++ media-player/srcs/rendering/texture.cxx
@@ -12,7 +12,8 @@ void Texture::render(std::uint32_t x, std::uint32_t y) noexcept
     if (_fullscreen) {
         render(boost::optional<Rect>{}, boost::optional<Rect>{});
     } else {
-        render(Rect{ 0u, 0u, si.first, si.second }, Rect{ x, y, si.first, si.second });
+        render(Rect{ 0, 0, si.first, si.second },
+            Rect{ static_cast<std::int32_t>(x), static_cast<std::int32_t>(y), si.first, si.second });
     }
 }
